refactor: rewrite dragNdrop cache helpers
